gateware.fft: fix twiddle factors
The previous TFs matched those given in the paper
"The Fast Fourier Transform in Hardware: A Tutorial Based on an FPGA Implementation"
but the overall FFT results were incorrect (mirrored around the real
axis).

This changes the TF equation to match other literature and gives the
correct FFT result when compared to numpy.fft.